move get resources to its own group
